fix building error
